Move win32filestrean_tests.cxx contents into fileappender.cxx.
